Tests: move AbsolutePath tests into single class
