+++ TruePath.Tests/AbsolutePathTests.cs
@@ -6,7 +6,7 @@ using System.Diagnostics;
 
 namespace TruePath.Tests;
 
-public partial class AbsolutePathTests
+public class AbsolutePathTests
 {
     [Fact]
     public void ConstructionTest()
@@ -226,6 +226,84 @@ public partial class AbsolutePathTests
         var incorrectCaseDirectory = newDirectory.Parent!.Value / "FOOBAR";
         var result = incorrectCaseDirectory.Canonicalize();
         Assert.Equal(newDirectory.Value, result.Value);
+    }
+
+    [Fact]
+    public void EqualsUseStrictStringPathComparer_SamePaths_True()
+    {
+        // Arrange
+        var currentDirectory = Environment.CurrentDirectory;
+        var nonCanonicalPath = currentDirectory;
+
+        var path1 = new AbsolutePath(currentDirectory);
+        var path2 = new AbsolutePath(nonCanonicalPath);
+
+        // Act
+        var equals = path1.Equals(path2, StrictStringPathComparer.Comparer);
+
+        // Assert
+        Assert.True(equals);
+    }
+
+    [Fact]
+    public void EqualsUseStrictStringPathComparer_NotSamePaths_False()
+    {
+        // Arrange
+        var currentDirectory = Environment.CurrentDirectory;
+        var nonCanonicalPath = new string(currentDirectory.MakeNonCanonicalPath().ToArray());
+
+        var path1 = new AbsolutePath(currentDirectory);
+        var path2 = new AbsolutePath(nonCanonicalPath);
+
+        // Act
+        var equals = path1.Equals(path2, StrictStringPathComparer.Comparer);
+
+        // Assert
+        Assert.False(equals);
+    }
+
+    [Fact]
+    public void OnLinux_EqualsDefault_CaseSensitive_False()
+    {
+        // Arrange
+        if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows) || RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
+        {
+            return;
+        }
+
+        var currentDirectory = Environment.CurrentDirectory;
+        var nonCanonicalPath = new string(currentDirectory.MakeNonCanonicalPath().ToArray());
+
+        var path1 = new AbsolutePath(currentDirectory);
+        var path2 = new AbsolutePath(nonCanonicalPath);
+
+        // Act
+        var equals = path1.Equals(path2);
+
+        // Assert
+        Assert.False(equals);
+    }
+
+    [Fact]
+    public void OnWindowsOrOsx_EqualsDefault_CaseInsensitive_True()
+    {
+        // Arrange
+        if (!RuntimeInformation.IsOSPlatform(OSPlatform.Windows) && !RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
+        {
+            return;
+        }
+
+        var currentDirectory = Environment.CurrentDirectory;
+        var nonCanonicalPath = new string(currentDirectory.MakeNonCanonicalPath().ToArray());
+
+        var path1 = new AbsolutePath(currentDirectory);
+        var path2 = new AbsolutePath(nonCanonicalPath);
+
+        // Act
+        var equals = path1.Equals(path2);
+
+        // Assert
+        Assert.True(equals);
     }
 
     private static bool CreateJunction(string path, string target)
